app/api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
Flask==2.0.2
Flask-Admin==1.5.8
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
passlib==1.7.4
protobuf==3.19.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.16.0
Werkzeug==2.0.2
WTForms==3.0.1
